refactor(cli): extract runs subcommand into cli/commands/runs.mjs
Move the `runs` (list | diff | summary | digest) block out of main() into a
dedicated handler. Behavior, messages, and exit codes are unchanged; only the
enclosing function and relative import depth differ.
